+++ components/product/Overview.vue
@@ -44,6 +44,11 @@
       </CoreButton>
     </div>
   </ProductField>
+  <CoreButton v-if="product.affiliate" variant="custom" :href="product.affiliate" target="_blank"
+    class="flex items-center justify-center space-x-2 w-full border border-gray-300 text-gray-500 hover:text-black hover:bg-gray-100">
+    <span>{{ $t('common.affiliate-program') }}</span>
+    <Icon name="ic:outline-handshake" size="20" />
+  </CoreButton>
 </template>
 
 <script setup lang="ts">