+++ content/en/changelog.md
@@ -2,6 +2,11 @@
 
 All notable changes and updates made to this website
 
+## Apr 12, 2024
+
+- Minor tweaks to the product Overview section's layout
+- Added "Affiliate Program" information to the product
+
 ## Apr 7, 2024
 
 - Added website intelligence to the product Overview section